Enable development with Bundler
This commit makes aeolus-image ready for development when using Bundler.

Gemfile is added, just references the gemspec.

Gitignore file is added with Bundler-related things (ignores also
Gemfile.lock, which is common practice for gems).

Development dependencies, which are required to run specs but were
previously missing from the gemspec, are added.